+++ src/app/cart-item/models/cartItem.model.ts
@@ -1,0 +1,9 @@
+import {ProductModel} from '../../product/models/product.model';
+
+export class CartItemModel {
+  constructor(
+    public product: ProductModel,
+    public amount: number = 1
+  ) {
+  }
+}