+++ changelog.md
@@ -7,3 +7,12 @@
 6. Создала сервис `CartService`
 7. Добавила вывод итоговой суммы товаров корзины в компоненте `CartListComponent`
 8. Добавила фильтр на товары с флагом `isAvailable=false` в компоненте `ProductListComponent`
+
+### Task 2. Components
+1. Разбила приложение на модули
+2. `ProductComponent` сделала презентационным компонентом для `ProductListComponent`
+3. Реализовала возможность добавить товар в корзину, если он доступен
+4. Создала компонент `СartItemComponent`, который является презентационным для `CartListComponent`
+5. В `AppComponent` добавила шаблонную переменную `#appTitle`
+6. Добавила директиву `HighlightDirective`
+7. Товары в корзине группируются по имени